Modify config utils. (#726)
* Modify github's workflow

Remove job "check host". As the docker's
health check does the same thing. The
difference is that we do the check within
the container, to avoid troubles with
github's machine.

* Modify config utils

Config is now placed into boot.

Remove code related to cross site check.

Still a lot of dependency errors to resolve.