把首页轮换的图片放到数据库中 (#888)
* HomepageImage model

As a replacement for the hard-coded redirect.json

* Use HomepageImage for guide pictures

* Rename HomepageImage to GuidePicture

Because the code saves these images in a variable called "guidepics".

* Add management command to import existing redirect.json

* Revert "Rename HomepageImage to GuidePicture"

This reverts commit c0f8da8ce18a2323767e29238b7ff8c40e4d4f8d.

HomepageImage class represents a specific subset of images that will
appear on the homepage of a website.

* Add auxiliary fields for HomepageImage

* Add aux fields to import command

The problem is that we can't generate descriptions that don't make us sound like idiots.

* Use activated and sort_id fields in view function

* Admin site for HomepageImage

* Remove the import command

* Rename guidepics to homepage_image in code

* Add fallback for homepage image